feat(后台管理)：登录、商品管理（添加） refactor(all)：统一类型

+++ src/main/java/com/shijinglu/ecommerceresponsesystem/Dao/CategoryMapper.java
@@ -22,7 +22,7 @@ public interface CategoryMapper extends BaseMapper<Category> {
 
 
     @Select("SELECT category_id FROM category WHERE category_name = #{categoryName}")
-    Long getCategoryIdByName(String categoryName);
+    Integer getCategoryIdByName(String categoryName);
 
 
 }
